# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Insta API",
        "description": "API документация для проекта Insta",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user profile",
                "description": "Returns detailed profile information of a user",
                "operationId": "da465f43567478f255c1892186f2af9d",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile data"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{user}/posts": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user's posts",
                "description": "Returns a list of posts created by a user",
                "operationId": "c6edc1bff02524861905ae1a9b73be65",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user's posts"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{user}/subscriptions": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user's subscriptions",
                "description": "Returns a list of users the specified user is subscribed to",
                "operationId": "c88820799e88a4d5ece265c44efdd13d",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of subscriptions"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{user}/subscribers": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user's subscribers",
                "description": "Returns a list of users who subscribed to the specified user",
                "operationId": "8f8ecb2d231d4aff9a8ebf1cb0f38a8a",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of subscribers"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{user}/subscribe": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Subscribe to a user",
                "description": "Subscribe to the specified user",
                "operationId": "3a55beac4b430e1f4285251c417198b2",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID to subscribe to",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Subscribed successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Paste your Sanctum access token here",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}